updated codebase

+++ utils.py
@@ -389,9 +389,9 @@ def load_vectors(data_loader,type='query'):
         return torch.from_numpy(data_loader()[1])
     elif type == 'answers':
         #answers are NN of the query points
-        return torch.from_numpy(data_loader()[2])
+        return torch.from_numpy(data_loader()[2].astype(np.int32))
     elif type == 'train':
-        return torch.from_numpy(data_loader()[0].astype(np.int32))
+        return torch.from_numpy(data_loader()[0])
     else:
         raise Exception('Unsupported data type')
